requirements.txt
asgiref==3.7.2
Django==4.2.2
django-modeltranslation==0.18.11
Pillow==9.5.0
psycopg2==2.9.6
python-dotenv==1.0.1
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2023.3
